Remove all the tests about `maxInterStageShaderComponents`
This patch removes all the tests about `maxInterStageShaderComponents`
as it has been removed from the latest WebGPU SPEC.
